# static/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "The Unchained MHacks API",
        "version": "0.0.1",
        "title": "MHacks",
        "termsOfService": "https://mhacks.org",
        "contact": {
            "email": "me@antoniomika.me"
        },
        "license": {
            "name": "GNU General Public License v3.0",
            "url": "https://github.com/mhacks/mhacks-web/blob/master/LICENSE"
        }
    },
    "servers": [
        {
            "url": "https://mhacks.org/v1",
            "description": "Production"
        },
        {
            "url": "https://staging.mhacks.org/v1",
            "description": "Staging"
        },
        {
            "url": "http://localhost:3000/v1",
            "description": "Development"
        }
    ],
...
    "paths": {
...
        "/v1/application": {
            "get": {
                "tags": [
                    "Application"
                ],
                "summary": "",
                "description": "",
                "operationId": "getApplication",
                "responses": {
                    "200": {
                        "description": "Okay",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "description": "Whether or not the request was successful"
                                        },
                                        "application": {
                                            "$ref": "#/components/schemas/Application"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error (Unknown Error)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Application"
                ],
                "summary": "",
                "description": "",
                "operationId": "postApplication",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Application"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Okay",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "description": "Whether or not the request was successful"
                                        },
                                        "application": {
                                            "$ref": "#/components/schemas/Application"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error (Unknown Error)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/application/all": {
            "get": {
                "tags": [
                    "Application"
                ],
                "summary": "",
                "description": "",
                "operationId": "getApplicationAll",
                "responses": {
                    "200": {
                        "description": "Okay",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "description": "Whether or not the request was successful"
                                        },
                                        "application": {
                                            "$ref": "#/components/schemas/Application"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error (Unknown Error)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/application/confirm": {
            "get": {
                "tags": [
                    "Application"
                ],
                "summary": "",
                "description": "",
                "operationId": "getApplicationConfirm",
                "responses": {
                    "200": {
                        "description": "Okay",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "description": "Whether or not the request was successful"
                                        },
                                        "application": {
                                            "$ref": "#/components/schemas/Application"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error (Unknown Error)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Application"
                ],
                "summary": "",
                "description": "",
                "operationId": "postApplicationConfirm",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Application"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Okay",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "description": "Whether or not the request was successful"
                                        },
                                        "application": {
                                            "$ref": "#/components/schemas/Application"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error (Unknown Error)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/v1/mentor/application": {
            "get": {
                "tags": [
                    "MentorApplication"
                ],
                "summary": "",
                "description": "",
                "operationId": "getMentorApplication",
                "responses": {
                    "200": {
                        "description": "Okay",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "description": "Whether or not the request was successful"
                                        },
                                        "application": {
                                            "$ref": "#/components/schemas/MentorApplication"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error (Unknown Error)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "MentorApplication"
                ],
                "summary": "",
                "description": "",
                "operationId": "postMentorApplication",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MentorApplication"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Okay",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "description": "Whether or not the request was successful"
                                        },
                                        "application": {
                                            "$ref": "#/components/schemas/MentorApplication"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error (Unknown Error)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/v1/speaker/application": {
            "get": {
                "tags": [
                    "SpeakerApplication"
                ],
                "summary": "",
                "description": "",
                "operationId": "getSpeakerApplication",
                "responses": {
                    "200": {
                        "description": "Okay",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "description": "Whether or not the request was successful"
                                        },
                                        "application": {
                                            "$ref": "#/components/schemas/SpeakerApplication"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error (Unknown Error)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "SpeakerApplication"
                ],
                "summary": "",
                "description": "",
                "operationId": "postSpeakerApplication",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SpeakerApplication"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Okay",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "description": "Whether or not the request was successful"
                                        },
                                        "application": {
                                            "$ref": "#/components/schemas/SpeakerApplication"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error (Unknown Error)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "components": {
        "schemas": {
...
            "Application": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "The id of the document"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "general_header": {},
                    "user": {
                        "type": "string",
                        "description": "Email"
                    },
                    "birthday_ts": {
                        "type": "number",
                        "format": "integer",
                        "description": "Date of Birth"
                    },
                    "birthday": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Date of Birth"
                    },
                    "university": {
                        "type": "string",
                        "description": "University"
                    },
                    "major": {
                        "type": "string",
                        "description": "Major"
                    },
                    "tshirt": {
                        "type": "string",
                        "description": "T-Shirt",
                        "enum": [
                            "unselected",
                            "xs",
                            "s",
                            "m",
                            "l",
                            "xl",
                            "2xl",
                            "3xl"
                        ]
                    },
                    "experience": {
                        "type": "string",
                        "description": "Experience",
                        "enum": [
                            "novice",
                            "experienced",
                            "veteran"
                        ]
                    },
                    "links_header": {},
                    "resume": {
                        "type": "string",
                        "format": "binary",
                        "description": "Resume"
                    },
                    "github": {
                        "type": "string",
                        "description": "GitHub"
                    },
                    "linkedin": {
                        "type": "string",
                        "description": "LinkedIn"
                    },
                    "devpost": {
                        "type": "string",
                        "description": "DevPost"
                    },
                    "portfolio": {
                        "type": "string",
                        "description": "Portfolio"
                    },
                    "demographics_header": {},
                    "race": {
                        "type": "string",
                        "description": "Race",
                        "enum": [
                            "unselected",
                            "white",
                            "black",
                            "am-indian-alaskan",
                            "asian",
                            "hispanic",
                            "other",
                            "prefer-not"
                        ]
                    },
                    "sex": {
                        "type": "string",
                        "description": "Sex",
                        "enum": [
                            "unselected",
                            "male",
                            "female",
                            "non-binary",
                            "prefer-not"
                        ]
                    },
                    "short_answer_headers": {},
                    "why_mhacks": {
                        "type": "string",
                        "description": "Why do you want to come to MHacks?"
                    },
                    "favorite_memory": {
                        "type": "string",
                        "description": "What is your favorite memory from  MHacks (if applicable)?"
                    },
                    "anything_else": {
                        "type": "string",
                        "description": "Anything else you would like to tell us?"
                    },
                    "needs_reimbursement": {
                        "type": "boolean",
                        "description": "Do you need travel reimbursement?"
                    },
                    "departing_from": {
                        "type": "string",
                        "description": "Departing From"
                    },
                    "requested_reimbursement": {
                        "type": "number",
                        "description": "How much reimbursement do you expect to need?"
                    },
                    "status": {
                        "type": "string",
                        "description": "Status",
                        "default": "unread",
                        "enum": [
                            "unread",
                            "waitlisted",
                            "accepted"
                        ]
                    },
                    "score": {
                        "type": "number",
                        "description": "Score"
                    },
                    "reader": {
                        "type": "string",
                        "description": "Reader"
                    },
                    "reimbursement": {
                        "type": "number",
                        "description": "Reimbursement"
                    },
                    "reader_filter": {},
                    "reader_schema": {}
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}